+++ src/api.ts
@@ -13,7 +13,7 @@ import {NextAfterTokenWalker} from "./parse/grammar/interpreter"
 let API:any = {}
 
 // semantic version
-API.VERSION = "0.10.1"
+API.VERSION = "0.10.2"
 
 // runtime API
 API.Parser = Parser